fix(runtime): make main pass just clippy again
Why: fifteen clippy errors (unnecessary_cast on already-typed ABI
constants, manual_noop_waker, unnecessary_get_then_check) in the test
modules of dispatch/net.rs and vcpu_loop/continuation.rs landed without
the gate; `just clippy` on unmodified main exited 101, so every branch
built on it reported a red gate that was not its own.

What: drop the same-type casts, use `Waker::noop()` for the plain
no-op waker, keep the two `DropCheckingWake` impls (the tests observe
WHEN that waker drops, which `Waker::noop()` cannot) behind a per-site
allow with the reason, and express the two absence assertions as
`!contains_key`.

Verified: `just clippy` exits 0; the touched continuation tests pass
under `RUST_TEST_THREADS=1 cargo test -p carrick-runtime --lib`.

+++ crates/carrick-runtime/src/dispatch/net.rs
@@ -11712,7 +11712,7 @@ mod nested_epoll_readiness_tests {
         let pollfds_addr = 0x1100u64;
         let pfd = LinuxPollFd {
             fd: outer,
-            events: LINUX_POLLIN as i16,
+            events: LINUX_POLLIN,
             revents: 0,
         };
         guest_mem
@@ -11791,7 +11791,7 @@ mod nested_epoll_readiness_tests {
 
         assert_eq!(ppoll_ready, DispatchOutcome::Returned { value: 1 });
         let out_pfd: LinuxPollFd = read_kernel_struct(&guest_mem, pollfds_addr).unwrap();
-        assert_eq!(out_pfd.revents & (LINUX_POLLIN as i16), LINUX_POLLIN as i16);
+        assert_eq!(out_pfd.revents & LINUX_POLLIN, LINUX_POLLIN);
     }
 
     #[test]
